+++ FontEnd/erp/src/app/modules/human-resources/configuration/ward/ward.service.ts
@@ -3,6 +3,8 @@ import { Injectable } from '@angular/core';
 import { Observable, of } from 'rxjs';
 import { switchMap } from 'rxjs/operators';
 
+import { SessionContext } from 'src/app/core/session.context';
+import { PermissionViewModel } from './../../../../core/models/permission.model';
 import { PagingModel } from './../../../../core/models/paging.model';
 import { DialogService } from './../../../../core/services/dialog.service';
 import { ApiService } from './../../../../core/services/api.service';
@@ -15,6 +17,9 @@ import { FilterModel } from 'src/app/core/models/filter-table.model';
 @Injectable()
 export class WardService {
 
+  permission = new PermissionViewModel();
+  moduleName = 'HR';
+  functionCode = 'HR_CONF_WARD';
   url = {
     list: APIUrlConstants.commonApi + 'ward/get-list',
     dropdown: APIUrlConstants.commonApi + 'ward/dropdown',
@@ -26,7 +31,13 @@ export class WardService {
 
   constructor(
     private api: ApiService,
-    private dialogService: DialogService) { }
+    private dialogService: DialogService,
+    private context: SessionContext) { }
+
+  getPermission(): PermissionViewModel {
+    this.permission = this.context.getPermissionByForm(this.moduleName, this.functionCode);
+    return this.permission;
+  }
 
   getList(paging: PagingModel, searchText: string) {
     const filter = new FilterModel();
@@ -46,6 +57,9 @@ export class WardService {
   }
 
   save(model: WardViewModel, action: FormActionStatus): Observable<ResponseModel> {
+    if (this.permission.allowInsert === false && this.permission.allowUpdate === false) {
+      return;
+    }
     switch (action) {
       case FormActionStatus.Insert:
         return this.api.insert(this.url.insert, model);
@@ -55,6 +69,9 @@ export class WardService {
   }
 
   confirmDelete(itemId: number, version: any): Observable<ResponseModel> {
+    if (this.permission.allowDelete === false) {
+      return;
+    }
     return this.dialogService.openConfirmDeleteDialog().pipe(
       switchMap((confirmResponse: boolean) => {
         if (confirmResponse === true) {
@@ -67,6 +84,9 @@ export class WardService {
   }
 
   delete(itemId: number, version: any): Observable<ResponseModel> {
+    if (this.permission.allowDelete === false) {
+      return;
+    }
     return this.api.delete(this.url.delete, {id: itemId, rowVersion: version });
   }
 }
